arcade(invaders): drop SP folder, rename MP → canonical /invaders/
Phase J (#593) shipped MP feature parity + a 301 from /invaders → /invaders-mp/. With MP proven stable, this collapses the two folders into one:

- Delete docs/arcade/invaders/ (SP HTML + 4 orphaned bgm-*.mp3, ~7.4 MB)
- git mv docs/arcade/invaders-mp → docs/arcade/invaders
- sfx-shoot.mp3 + sfx-kill.wav move with the rename (were already cross-loaded by MP); index.html switches to ./ paths
- Worker.ts: drop the /invaders → /invaders-mp/ 301; room-code rewrite regex retargeted to /invaders/<segment>; legacy /invaders-mp/* and /invaders-2p/* both 301 to /invaders/* (preserves any in-flight QR-code links)
- wrangler.toml: drop run_worker_first (no asset overrides needed now)
- room.ts: import path follows the rename; v17/v19 history comments left as-is
- localStorage keys keep the legacy 'invaders-mp.' prefix to preserve existing players' saved names + avatar salts

Cabinet tile already pointed at 'invaders/' (it had been redirecting via Phase J) — works directly now.
