+++ docs/arcade/invaders/sprites.js
