+++ docs/plans/arcade-framework-roadmap.md
@@ -38,7 +38,7 @@ us toward an author (human or agent) only ever writing those four things?
 
 ## Why MP-as-default is viable (not aspirational)
 
-`invaders-mp` already proves it. It is **host-authoritative**: `engine.js`
+`invaders` already proves it. It is **host-authoritative**: `engine.js`
 (~1086 LoC, ES module, typed via `engine.d.ts`) is the sim, and the *same code*
 runs in the client's host-mode and in the Cloudflare DO worker. As of **#589**
 (`solo p1 skips the DO`), a solo player runs that sim **purely locally at 0 RTT**
@@ -70,9 +70,9 @@ SP and MP do **not** compete over the sim or the chrome. They differ only in the
 | Block | Where | Status |
 |---|---|---|
 | Cabinet chrome — splash/attract, initials, leaderboard, pause, end-overlay, touch, audio, iframe-host, cabinet/overlays/pixel-font/touch/splash CSS | `docs/arcade/shared/` | ✅ Shared, good APIs, headless tests for touch + splash |
-| Authoritative sim pattern (`engine.js` + `.d.ts`) | `docs/arcade/invaders-mp/` | 🟡 Exists for one game; not generalized/documented as a contract |
-| Pixel-render primitives (`sprites.js`) | `docs/arcade/invaders-mp/` | 🟡 invaders-only |
-| **Session / transport** — DO + WebRTC + lobby + seats + host harness + solo-skips-DO | inline in `invaders-mp/index.html` + `infra/oyster-arcade-mp/` | 🔴 Works, but **locked inline** — not a reusable block |
+| Authoritative sim pattern (`engine.js` + `.d.ts`) | `docs/arcade/invaders/` | 🟡 Exists for one game; not generalized/documented as a contract |
+| Pixel-render primitives (`sprites.js`) | `docs/arcade/invaders/` | 🟡 invaders-only |
+| **Session / transport** — DO + WebRTC + lobby + seats + host harness + solo-skips-DO | inline in `invaders/index.html` + `infra/oyster-arcade-mp/` | 🔴 Works, but **locked inline** — not a reusable block |
 | **Cabinet bootstrap / new-game scaffold** (`Arcade.Game.boot(config)`) | — | 🔴 Doesn't exist; every game hand-wires DOM + `mount()` calls |
 | `Arcade.Music` (BGM control) | `docs/arcade/shared/music.js` | ✅ Shipped (#594) — multiplexer + autoplay-retry + gain×slider compose. invaders + space-jumper migrated; rocket-ship in the SP-0 PR. |
 
@@ -89,18 +89,19 @@ not strict dependency.
   autoplay-retry) into a shared module. The only *systematic* duplication left
   across the SP games. Model it on invaders' `playBgm(which)` multiplexer.
   Small, standalone. Unblocks Rocket Ship's last non-adopted bit.
-- **`MP-1` — invaders-mp → SP parity.** The existing phased plan
-  (`docs/plans/invaders-mp-sp-audit.md`, Phases D–J): lives, per-player score,
-  shields, combos, supers, stages, bosses, then adopt chrome + leaderboard and
-  deprecate `/invaders/`. The proving ground that hardens `engine.js` / `room.ts`.
+- **`MP-1` — invaders MP → SP parity.** ✅ Shipped. The phased plan
+  (`docs/plans/invaders-mp-sp-audit.md`, Phases D–J) reached J: lives, per-player
+  score, shields, combos, supers, stages, bosses, chrome + leaderboard. SP folder
+  deleted; the MP build is now the canonical `/invaders/`. Hardened `engine.js`
+  / `room.ts`.
 - **`FW-1` — Extract the session/transport block** (`Arcade.Session` /
-  `Arcade.Net`) from invaders-mp into `shared/`. The keystone for "spin up a new
+  `Arcade.Net`) from invaders into `shared/`. The keystone for "spin up a new
   MP game." **Sequenced after `MP-1`** so it's extracted from stable, proven code
   (the 2026-05-18 spec's "extract once the duplication is stable" rule — which
   was right).
 - **`FW-2a` — Documented sim contract.** The conceptual, low-risk half: pin down
   the `engine.js` interface (`init`/`step`, input shape, state shape, determinism
-  expectations) as a written contract + `.d.ts`, derived from invaders-mp's
+  expectations) as a written contract + `.d.ts`, derived from invaders'
   proven engine. No DOM, no lifecycle — just the shape an author targets.
 - **`FW-2b` — Cabinet bootstrap** (`Arcade.Game.boot(config)`). The larger half:
   a bootstrap that injects the standard DOM and wires chrome + session + sim +
